<Update>: Split code into CopyMove.cs and Compress.cs
